memblock: extend memblock_cap_memory_range to multiple ranges
The memblock_cap_memory_range() removes all the memory except the range
passed to it. Extend this function to recieve an array of memblock_regions
that should be kept. This allows switching to simple iteration over
memblock arrays with 'for_each_mem_range_rev' to remove the unneeded memory.

Enable use of this function in arm64 for reservation of multile regions for
the crash kernel.

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>