Allocate page from netns mm during mount
When the socket file system is mounted, it allocates an inode. I don't know
why sockets is a file system, and I don't know why it needs an inode, but
I do know that it must be allocated from the correct address space, which
in this case belongs to the netns.

Signed-off-by: Joel Nider <joeln@il.ibm.com>