replicate the superblock for sockets
There must be a 1:1 relationship between the superblock and the kmem_cache
it uses for allocating inodes. Therefore, I have moved the cache object to
be a field in the superblock to guarantee this relationship. Most file
systems ignore the superblock in alloc_inode() and allocate directly from
the static cache object. Now, socket code allocates the inode from the
cache in the superblock. In order to guarantee that the the correct cache
is being used (remember, there is now a private cache per network namespace)
I am creating a superblock per network namespace. Only the first one is
ever mounted (init_ns), and the others are just containers for the cache.
Also, there is still the concept of a cache owner, which can contain a
selection of callbacks to be called on events such as page allocation from
the owner's cache. This gives us a chance to act on the event, for example
mapping or unmapping the page in the other network namespaces to prevent
unauthorized access from outside the owning namespace.

Signed-off-by: Joel Nider <joeln@il.ibm.com>