Use process page tables even when in netns
We overwrite the kernel entries of the processes' page tables
when entering the netns. That way, each process keeps its own context,
even across context switches. It also means multiple processes can
use the namespace simultaneously. It also means there is no problem
entering the namespace recursively.

This patch also includes some fixes to the socket superblock, which
was not getting the ops correctly, which prevented allocation of an
inode of the right type.

Signed-off-by: Joel Nider <joeln@il.ibm.com>