x86/mm/kaslr: refactor to use enum indices for regions
The KASLR randomization code currently refers to specific regions, such
as the vmalloc area, by literal indices into an array. When adding new
regions, we have to be careful to also change all indices that may
potentially change. Avoid that risk by introducing an enum used as
indices.

[
original patch was
  signed by Marius Hillenbrand <mhillenb@amazon.de>
and cc'ed to
 Alexander Graf <graf@amazon.de>
 David Woodhouse <dwmw@amazon.co.uk>
]

Signed-off-by: Marius Hillenbrand <mhillenb@amazon.de>