x86/mm: add virtual memory region for exclusive mappings
The Linux kernel has a global address space that is the same for any
kernel code. This address space becomes a liability in a world with
processor information leak vulnerabilities, such as L1TF. With the right
cache load gadget, an attacker-controlled hyperthread pair can leak
arbitrary data via L1TF. Disabling hyperthreading is one recommended
mitigation, but it comes with a large performance hit for a wide range
of workloads.

An alternative mitigation is to not make certain data in the kernel
globally visible, but only when the kernel executes in the context of
the process where this data belongs to.

Introduce a region for process-exclusive memory mappings into the kernel's
virtual address space. When enabled, the region occupies the upper half of
the direct map of the physical memory and the later is cut in half.

[
rppt:
 * renamed proclocal to exclusive
 * extended the virtual address range to cover the entire physical memory

original patch was
 signed by Marius Hillenbrand <mhillenb@amazon.de>
 inspired by: Julian Stecklina <js@alien8.de> (while jsteckli@amazon.de)
and cc'ed to
 Alexander Graf <graf@amazon.de>
 David Woodhouse <dwmw@amazon.co.uk>
]

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>