mm: add support for kernel exclusive pages and __GFP_EXCLUSIVE allocations
The kernel exclusive pages are mapped in the dedicated part of the process
address space and dropped from the normal direct map.

The pages allocated with __GFP_EXCLUSIVE become "kernel exclusive" and as
such they are only visible in the context of the owning process.

On exit of this process the exclusive pages are added back to the kernel
direct map and the page table entries covering the exclusive are are teared
down.

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>