dev: exit: use ns_pgd rather than init_mm after mm is torn down
During exit() the mm of the process is freed. When this happens, the
context is switched to the &init_mm which wouldn't have mappings of the
exclusive pages used for e.g. socket caches and this will cause crashes
when the objects that use the exclusive pages are freed.
Temporarily switch to mm->ns_pgd after the process mm is destroyed and then
use &init_mm in the end of do_exit().

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>