KVM: Handle secretmem
Secretmem is not mapped into kernel direct mapping and requires special
handling if KVM wants to use it as guest memory.

The feature is gated by CONFIG_HAVE_KVM_PROTECTED_MEMORY and requires
an architecture to provide ioremap_cache_force().

kvm_share_memory() is impelementation of KVM_HC_MEM_SHARE. It replaces
secretmem in a given range with anonymous mapping which can be used for
guest-host communication.

Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>