x86/setup: don't remove E820_TYPE_RAM for pfn 0
The first 4Kb of memory is a BIOS owned area and to avoid its allocation
for the kernel it was not listed in e820 tables as memory.  As the result,
pfn 0 was never recognised by the generic memory management and it is not
a part of neither node 0 nor ZONE_DMA.

If set_pfnblock_flags_mask() would be ever called for the pageblock
corresponding to the first 2Mbytes of memory, having pfn 0 outside of
ZONE_DMA would trigger

	VM_BUG_ON_PAGE(!zone_spans_pfn(page_zone(page), pfn), page);

The beginning of the memory is always reserved with memblock during
setup_arch().  This reservation is enough to ensure that the kernel does
not touch the BIOS area and it is not necessary to remove E820_TYPE_RAM for
pfn 0.

The only corner case that presumed that pfn 0 is IORESOURCE_MEM rather than
IORESOURCE_SYSTEM_RAM is devmem_is_allowed() that granted access to the
first 1M in iomem range and disallowed access in that range for System RAM.
The ability to read the BIOS setup data stored in the first memory page is
required by legacy software, for instance LILO.

Remove the update of e820 table that changes the type of pfn 0 and waive the
restriction to access the first 4Kb of System RAM in devmem_is_allowed().

Link: https://www.linuxquestions.org/questions/slackware-14/slackware-current-lilo-vesa-warnings-after-recent-updates-4175689617/#post6214439
Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>