asi/slub: add support for kmem caches with __GFP_EXCLUSIVE
Add support for __GFP_EXCLUSIVE to kmem_cache_alloc() and kmalloc()
families.

Whenever there is a kmem_cache_alloc() or kmalloc() request with
__GFP_EXCLUSIVE set, a dedicated cache bound to the current ASI context is
used to serve this request. Such caches are lazily allocated when there is
the first allocation request with __GFP_EXCLUSIVE for a particular kmem
cache.

The pages that constitute the slabs are allocated with __GFP_UNMAP rather
than __GFP_EXCLUSIVE because it is enough to drop them from the direct map
and then slub code will take care of their mapping into the appropriate
virtual addresses.

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>