arm: switch to generic memblock_setup_resources()
The registration of "System RAM" resources on arm is very similar to the
generic version.

The major differences are the registration of "System RAM (boot alias)"
areas and minor differences in the way kernel image resources are reserved.

Replace the open coded registration of the "System RAM" resource and the
requests for kernel image and data areas in the resource tree with the
generic implementation.

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>