mm/page_alloc: introduce __GFP_UNMAPPED and MIGRATE_UNMAPPED
When set_memory or set_direct_map APIs used to change attribute or
permissions for chunks of several pages, the large PMD that maps these
pages in the direct map must be split. Fragmenting the direct map in such
manner causes TLB pressure and, eventually, performance degradation.

To avoid excessive direct map fragmentation, add ability to allocate
"unmapped" pages with __GFP_UNMAPPED flag and a new migrate type
MIGRATE_UNMAPPED that will contain pages that are mapped with PTEs in the
direct map.

When allocation flags include __GFP_UNMAPPED, the allocated page(s) are
remapped at PTE level in the direct map along with the other free pages in
the same pageblock. The free pages in that pageblock are placed on
MIGRATE_UNMAPPED free lists and subsequent __GFP_UNMAPPED request are
served from the MIGRATE_UNMAPPED free lists.

This way, a single split of a large PMD creates a cache of PTE-mapped pages
that can be used without the need to split another PMD.

The pages are removed from the direct map late during allocation to keep
the existing logic in prep_new_page() for clearing, poisoning and other
page content accesses.

When a page from MIGRATE_UNMAPPED pageblock is freed, it is restored in the
direct map to allow proper page contents access in free_pages_prepare().

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>