x86/cet/shst: allow address specification for map_shadow_stack()
For checkpoint/restore of tasks with shadow stack enabled it is
necessary to recreate the shadow stack VMAs at the same addresses they
were at the time of the checkpoint.

Add address parameter to map_shadow_stack() to support this. When a
non-zero address is passed to map_shadow_stack(), it passes this address
to mmap() and sets MAP_FIXED in the mmap() flags.

Signed-off-by: Mike Rapoport <rppt@linux.ibm.com>